pyproject.toml
[tool.poetry]
name = "py-objectstore"
version = "0.2.3"
description = "Simplified API towards object stores"
authors = ["Per Buer <per.buer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
apache-libcloud = "^2.8.0"
lockfile = "^0.12.2"
cryptography = "^2.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest_dependency = "^0.5.1"
pytest-cov = "^2.8.1"
pylint = "^2.4.4"
python-dotenv = "^0.12.0"
bump2version = "^1.0.0"
twine = "^3.1.1"
keyring = "^21.2.0"
artifacts-keyring = "^0.2.9"
pytest-azurepipelines = "^0.8.0"
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
